services/python/fastapi/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.7
pydantic==2.0.0
python-dotenv==1.0.0
